fix(charts): 统计曲线按 0 轴分色、百分比轴停在 0..100、盈亏比保留两位小数
用户报的四条一起修（`tr-ui` 的折线图组件 + 股票统计页）：

- 面积按 0 轴分色：累计盈亏 / 期望值这类跨零曲线原来整块一个颜色，盈利段看着像亏损段。现在面积、折线、数据点三处按符号上色（红涨绿跌），交点插值落在 0 轴上；SVG 定点改写收敛成一个标签扫描器（填充基线 / 实心点 / 分色共用）。多序列图表不受影响（`sign_colors` 只在单序列生效）。
- 胜率 Y 轴停在 0..100：百分比类指标有天生边界，而自算刻度要求"上界严格大于数据最大值"，数据正好压在 100%（首笔就 100）时会多撑出一档 150%。新增 `ChartConfig::y_bounds`（有负数时 -100..100），上界贴着数据时只抬一个刻度文字里看不见的微量。最大回撤同样是百分比但没有天花板（4% 画在 0..100 上没法看），仍按数据挑；它也不是盈亏，保持单色绿、与自己那张 KPI 卡片同色。
- 盈亏比不再被取整成 1：曲线取值从 i64 改成 f64（金额仍是分），比值走新的 `ChartValueKind::Ratio`（两位小数）—— 之前 0.81 被 round 成 1，曲线是条直线；百分比也顺带修好（tooltip 83.33% 不再变 83.00%）。
- 跨零不再强制上下对称：-35..90 原来给 ±100（底下 50 的空白里一个点都没有），现在各按数据贴边（-50..100），0 仍落在刻度上。

回归：`fixtures/chart-tests.ps1` 从 7 条扩到 11 条（新增百分比上下界、盈亏比两位小数、跨零贴边、0 轴分色 —— 后者在 charts-rs 1.0.0 的真实 SVG 上逐字节断言）；e2e `chart-tests` / `ui-shots` / `ui-crud` / `ui-drag` / `ui-stock` 全绿（真启动 + UIA 驱动），并用一次性截图脚本人工核过五个指标的曲线。

+++ crates/tr-ui/src/pages/data_analysis.rs
@@ -776,7 +776,10 @@ fn chart_panel(
                                 ChartSeries::new(
                                     line.label.clone(),
                                     series_color(&transaction_type, index),
-                                    line.data.iter().map(|point| point.amount).collect(),
+                                    line.data
+                                        .iter()
+                                        .map(|point| point.amount as f64)
+                                        .collect(),
                                 )
                             })
                             .collect::<Vec<_>>();
